Improved buffering for streamed content
Improved buffering, BoxConn will now wait a short time (10 microsecs) for additional .Write calls to fill up a frame
this will be skipped if the frame is already full (meaning to delays for larger blocks of data) or if there is already
additional data available to be written.
Project structure is also improved.